WEBAR/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
Pillow==9.0.1
requests==2.27.1
cryptography>=3.4